+++ Stagehand/Utils/DataTransferFragment.cs
@@ -1,4 +1,6 @@
 using Stagehand.AssetLibrary;
+using Stagehand.Definitions.Serialization;
+using Stagehand.Editor.DefinitionEditors.Objects;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -15,25 +17,45 @@ namespace Stagehand.Utils;
 /// </summary>
 [JsonPolymorphic(TypeDiscriminatorPropertyName = "Type")]
 [JsonDerivedType(typeof(AssetBookmarkService.BookmarkDataTransferFragment), typeDiscriminator: "BookmarkDataTransferFragment")]
+[JsonDerivedType(typeof(ObjectDefinitionDataTransferFragment), typeDiscriminator: "ObjectDefinitionDataTransferFragment")]
 public abstract record class DataTransferFragment()
 {
+    private static readonly JsonSerializerOptions _dataTransferFramentJsonOptions = new()
+    {
+        WriteIndented = true,
+        Converters =
+        {
+            new JsonStringEnumConverter(),
+            new Vector2JsonConverter(),
+            new Vector3JsonConverter(),
+            new Vector4JsonConverter(),
+        },
+    };
+
     public string ToDataString()
     {
-        return JsonSerializer.Serialize(this);
+        return JsonSerializer.Serialize(this, _dataTransferFramentJsonOptions);
     }
 
     public async Task WriteToStream(Stream utf8JsonStream, CancellationToken cancellationToken)
     {
-        await JsonSerializer.SerializeAsync(utf8JsonStream, cancellationToken).ConfigureAwait(false);
+        await JsonSerializer.SerializeAsync(utf8JsonStream, this, _dataTransferFramentJsonOptions, cancellationToken).ConfigureAwait(false);
     }
 
     public static DataTransferFragment? FromDataString(string dataString)
     {
-        return JsonSerializer.Deserialize<DataTransferFragment>(dataString);
+        try
+        {
+            return JsonSerializer.Deserialize<DataTransferFragment>(dataString, _dataTransferFramentJsonOptions);
+        }
+        catch (JsonException)
+        {
+            return null;
+        }
     }
 
     public static async Task<DataTransferFragment?> FromStream(Stream utf8JsonStream, CancellationToken cancellationToken)
     {
-        return await JsonSerializer.DeserializeAsync<DataTransferFragment>(utf8JsonStream, cancellationToken: cancellationToken).ConfigureAwait(false);
+        return await JsonSerializer.DeserializeAsync<DataTransferFragment>(utf8JsonStream, _dataTransferFramentJsonOptions, cancellationToken: cancellationToken).ConfigureAwait(false);
     }
 }